metadocs/models/user/requirements.txt
fastapi
deta
bcrypt
python-dotenv
jinja2
aiofiles
python-multipart